+++ extension.js
@@ -1,44 +1,74 @@
 // The module 'vscode' contains the VS Code extensibility API
 // Import the module and reference it with the alias vscode in your code below
-const fs = require('fs');
-const path = require('path');
 const vscode = require('vscode');
-const FsmContentProvider = require('./lib/content-provider');
-const fetchMedia = require('./lib/media');
+const getWebviewContent = require('./lib/content');
 
 // this method is called when your extension is activated
 // your extension is activated the very first time the command is executed
 function activate(context) {
-  const registerCommand = vscode.commands.registerCommand;
-  const previewUri = vscode.Uri.parse('fsm-viewer:authority/view');
+  context.subscriptions.push(
+    vscode.commands.registerCommand('fsmViewer.view', () => {
+      const _disposables = [];
 
-  const provider = new FsmContentProvider(context);
+      const panel = vscode.window.createWebviewPanel(
+        'fsmViewer',
+        'FSM Viewer',
+        vscode.ViewColumn.Two,
+        {
+          // Enable scripts in the webview
+          enableScripts: true
+        }
+      );
 
-  const registration = vscode.workspace.registerTextDocumentContentProvider('fsm-viewer', provider);
+      vscode.workspace.onDidChangeTextDocument(
+        e => {
+          if (e.document === vscode.window.activeTextEditor.document) {
+            panel.webview.html = getWebviewContent(context);
+          }
+        },
+        null,
+        _disposables
+      );
 
-  const view = registerCommand('fsm-viewer.view', () => {
-    vscode.commands
-    .executeCommand('vscode.previewHtml', previewUri, vscode.ViewColumn.Two, 'FSM Viewer')
-    .then(success => {}, reason => vscode.window.showErrorMessage(reason));
-  });
+      vscode.window.onDidChangeTextEditorSelection(
+        e => {
+          if (e.textEditor === vscode.window.activeTextEditor) {
+            panel.webview.html = getWebviewContent(context);
+          }
+        },
+        null,
+        _disposables
+      );
 
-  vscode.workspace.onDidChangeTextDocument(e => {
-    if (e.document === vscode.window.activeTextEditor.document) {
-      provider.update(previewUri);
-    }
-  });
+      vscode.window.onDidChangeActiveTextEditor(
+        e => {
+          console.log(e);
+          // panel.webview.html = getWebviewContent(context);
+        },
+        null,
+        _disposables
+      );
 
-  vscode.window.onDidChangeTextEditorSelection(e => {
-    if (e.textEditor === vscode.window.activeTextEditor) {
-      provider.update(previewUri);
-    }
-  });
+      panel.onDidDispose(
+        () => {
+          console.log('panel closed');
 
-  vscode.window.onDidChangeActiveTextEditor(e => {
-    console.log(vscode.window.activeTextEditor.document.uri === previewUri);
-    provider.update(previewUri);
-  });
+          while (_disposables.length) {
+            const item = _disposables.pop();
+            if (item) {
+              item.dispose();
+            }
+          }
+        },
+        null,
+        context.subscriptions
+      );
 
+      panel.webview.html = getWebviewContent(context);
+    })
+  );
+
+  /*
   const save = registerCommand('fsm-viewer.save', () => {
     fetchMedia({
       format: 'svg',
@@ -48,16 +78,21 @@ function activate(context) {
         if (err) {
           vscode.window.showErrorMessage('SVG export failed');
         } else {
-          vscode.window.showInputBox({
-            prompt: 'Relative to project root',
-            placeHolder: 'Type the output file name'
-          }).then((input) => {
-            try {
-              fs.writeFileSync(path.join(vscode.workspace.rootPath, input), body);
-            } catch (e) {
-              console.log('not a file');
-            }
-          });
+          vscode.window
+            .showInputBox({
+              prompt: 'Relative to project root',
+              placeHolder: 'Type the output file name'
+            })
+            .then(input => {
+              try {
+                fs.writeFileSync(
+                  path.join(vscode.workspace.rootPath, input),
+                  body
+                );
+              } catch (e) {
+                console.log('not a file');
+              }
+            });
         }
       }
     }).then(mediaUrl => {
@@ -67,9 +102,9 @@ function activate(context) {
       });
     });
   });
-
-  context.subscriptions.push(view, save, registration);
+  */
 }
+
 exports.activate = activate;
 
 // this method is called when your extension is deactivated