+++ CHANGELOG.md
@@ -1,30 +1,44 @@
+## unpublished
+
+- Moved to Webview
+
 ## 1.7.0 - 2018-11-27
+
 - Dark theme support
 
 ## 1.6.0 - 2018-03-29
+
 - Handle FSMs with callbacks with quoted keys
 
 ## 1.5.0 - 2018-02-18
+
 - Handle FSMs without initial state
 
 ## 1.4.0 - 2018-02-16
+
 - Preview FSMs in files with import statements
 - Preview FSMs in JSX files
 
 ## 1.3.1 - 2018-01-30
+
 - Update dependencies
 
 ## 1.3.0 - 2018-01-18
+
 - Render state activities and guards
 
 ## 1.2.0 - 2018-01-17
+
 - Export diagram to SVG
 
 ## 1.1.0 - 2018-01-10
+
 - Support for events with multiple from states
 
 ## 1.0.1 - 2018-01-09
+
 - Support for arrow functions (via fsm2dot)
 
 ## 1.0.0 - 2018-01-08
+
 - Initial release